requirements.txt
asgiref==3.7.1
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.0
django-crispy-forms==1.12.0
django-filter==23.2
django-humanize==0.1.2
django-import-export==3.2.0
et-xmlfile==1.1.0
humanize==4.6.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
PyYAML==6.0
sqlparse==0.4.4
tablib==3.4.0
typing-extensions==4.6.1
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
 
whitenoise==6.5.0